// src/types/car.ts
export interface Car {
  id: string;
  name: string;
  brand: string;
  type: string;
  year: number;
  price: number;
  mileage: number;
  engine: string;
  horsepower: number;
  transmission: string;
  fuelType: string;
  color: string;
  interiorColor: string;
  features: string[];
  images: string[];
  description: string;
}